No longer exposing $context on matches
