+++ src/internal/SpecGrammar.ts
@@ -1,16 +1,15 @@
-import {Term} from "../Matchers";
-import {Concat} from "../matchers/Concat";
-import {RestOfInput} from "../matchers/skip/Skip";
-import {Break} from "../matchers/snobol/Break";
-import {Literal, Regex} from "../Primitives";
-import {Rep} from "../Rep";
+import { Concat } from "../matchers/Concat";
+import { RestOfInput } from "../matchers/skip/Skip";
+import { Break } from "../matchers/snobol/Break";
+import { Literal, Regex } from "../Primitives";
+import { Rep } from "../Rep";
 
 const elementReference = new Concat({
     $id: "component",
     _start: new Literal("${"),
     elementName: new Regex(/[a-zA-Z0-9_]+/),
     _end: new Literal("}"),
-} as Term);
+});
 
 export const specGrammar = new Concat({
     $id: "spec",
@@ -19,7 +18,7 @@ export const specGrammar = new Concat({
             $id: "literal, then component",
             literal: new Break(elementReference),
             element: elementReference,
-        } as Term)),
+        })),
     trailing: RestOfInput, //  matchEverything
 
 });
